app/python-microservice/requirements.txt

fastapi
uvicorn
requests
python-consul
